Sort the list of all movies
On the list of all movies page, the column headings for 'Movie Title' and 'Release Date' for a movie should be clickable links.  Clicking one of them should cause the list to be reloaded but sorted in ascending order on that column.  For example, clicking the 'release date' column heading should redisplay the list of movies with the earliest-released movies first; clicking the 'title' field should list the movies alphabetically by title.